[e2e] use testset.py to specify testset and testing more in e2e
